Mudancas para melhorar do desenvolvimento e passar em alguns da US4

+++ src/participacao/AlunoPosGraduando.java
@@ -13,7 +13,7 @@ public class AlunoPosGraduando extends Participacao {
 	}
 
 	@Override
-	public int calculaPontuacao() {
+	public double calculaPontuacao() {
 		return 0;
 	}
