+++ src/profissional/Gerente.java
@@ -10,7 +10,7 @@ public class Gerente extends Profissional {
 	}
 
 	@Override
-	public int calculaPontuacaoProfissional() {
+	public double calculaPontuacaoProfissional() {
 		int tempoRemunerado = (int) Math.floor(super.projeto.getDuracao() / 12);
 		int pontos = tempoRemunerado * 9;
 		return pontos;
