+++ src/profissional/Pesquisador.java
@@ -11,7 +11,7 @@ public class Pesquisador extends Profissional {
 	}
 
 	@Override
-	public int calculaPontuacaoProfissional() {
+	public double calculaPontuacaoProfissional() {
 		int tempoRemunerado = (int) Math.floor(super.projeto.getDuracao() / 12);
 		int pontos = tempoRemunerado * 6;
 		return pontos;
