+++ src/projeto/Projeto.java
@@ -162,8 +162,8 @@ public abstract class Projeto {
 		Collections.sort(participacoes);
 	}
 
-	public int calculaPontuacao() {
-		int acumulador = 0;
+	public double calculaPontuacao() {
+		double acumulador = 0;
 		for (Participacao participacao : participacoes) {
 			if (participacao.getClass() == AlunoGraduando.class) {
 				acumulador += 1;
